spelling fixes, removing unnecessary/empty lines, rename common-code folder to checks, update imports
