+++ src/controllers/update/due.diligence.overseas.entity.controller.ts
@@ -1,12 +1,16 @@
 import { NextFunction, Request, Response } from "express";
 
-import * as config from "../../config";
-import { getDueDeligence, postDueDeligence } from "../../utils/due.diligence";
+import { getDueDiligence, postDueDiligence } from "../../utils/due.diligence";
+import {
+  WHO_IS_MAKING_UPDATE_URL,
+  OVERSEAS_ENTITY_DUE_DILIGENCE_PAGE,
+  OVERSEAS_ENTITY_REVIEW_URL
+} from "../../config";
 
 export const get = (req: Request, res: Response, next: NextFunction) => {
-  getDueDeligence(req, res, next, config.WHO_IS_MAKING_UPDATE_URL, config.OVERSEAS_ENTITY_DUE_DILIGENCE_PAGE);
+  getDueDiligence(req, res, next, WHO_IS_MAKING_UPDATE_URL, OVERSEAS_ENTITY_DUE_DILIGENCE_PAGE);
 };
 
 export const post = (req: Request, res: Response, next: NextFunction) => {
-  postDueDeligence(req, res, next, config.OVERSEAS_ENTITY_REVIEW_URL, false);
+  postDueDiligence(req, res, next, OVERSEAS_ENTITY_REVIEW_URL, false);
 };
