+++ src/controllers/update/who.is.making.update.controller.ts
@@ -4,7 +4,7 @@ import {
   WHO_IS_MAKING_UPDATE_PAGE,
   OVERSEAS_ENTITY_PRESENTER_URL,
 } from "../../config";
-import { getWhoIsFilling, postWhoIsFilling } from "../../utils/who.is.making.filling";
+import { getWhoIsFilling, postWhoIsFilling } from "../../utils/who.is.making.filing";
 
 export const get = (req: Request, res: Response, next: NextFunction) => {
   getWhoIsFilling(req, res, next, WHO_IS_MAKING_UPDATE_PAGE, OVERSEAS_ENTITY_PRESENTER_URL);
