+++ src/middleware/navigation/checks/has.presenter.ts
