+++ src/middleware/navigation/update/has.presenter.middleware.ts
@@ -1,6 +1,6 @@
 import { Request, Response, NextFunction } from 'express';
 
-import { checkIfPresenterDataExists } from '../common-code/has.presenter';
+import { checkIfPresenterDataExists } from '../checks/has.presenter';
 import { UPDATE_LANDING_PAGE_URL } from '../../../config';
 
 export const hasUpdatePresenter = (req: Request, res: Response, next: NextFunction): void => {
