+++ views/includes/page-templates/due-diligence-overseas-entity.html
@@ -32,7 +32,6 @@
       {% include "includes/inset-text.html" %}
 
       {% include "includes/save-and-continue-button.html" %}
-
     </form>
   </div>
 </div>
